requirements.txt
biopython==1.81
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
click==8.1.3
docutils==0.17.1
Flask==2.2.2
gemmi==0.5.8
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
llvmlite==0.41.1
MarkupSafe==2.1.2
mmcif-pdbx==2.0.1
numba==0.58.1
numpy==1.23.5
pdb2pqr==3.5.2
Pillow==10.1.0
propka==3.5.0
pycparser==2.21
rdkit==2023.9.2
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
threadpoolctl==3.1.0
urllib3==1.26.14
Werkzeug==2.2.2

